Sync all models, DTOs, repositories and controllers to match Azure SQL schema exactly

+++ SkillBridgeTutors.API/SkillBridgeTutors.API/Interfaces/IDemoRepository.cs
@@ -5,9 +5,9 @@ namespace SkillBridgeTutors.API.Interfaces
     public interface IDemoRepository
     {
         Task<IEnumerable<DemoSlot>> GetAvailableSlotsAsync(int count = 5);
-        Task<DemoSlot?> GetSlotByIdAsync(int slotId);
+        Task<DemoSlot?> GetSlotByIdAsync(long slotId);
         Task<DemoBooking> CreateBookingAsync(DemoBooking booking);
-        Task<DemoBooking?> GetBookingByIdAsync(int bookingId);
+        Task<DemoBooking?> GetBookingByIdAsync(long bookingId);
         Task UpdateBookingAsync(DemoBooking booking);
         Task UpdateSlotAsync(DemoSlot slot);
     }
