+++ SkillBridgeTutors.API/SkillBridgeTutors.API/Interfaces/ILeadRepository.cs
@@ -5,7 +5,7 @@ namespace SkillBridgeTutors.API.Interfaces
     public interface ILeadRepository
     {
         Task<Lead> CreateAsync(Lead lead);
-        Task<Lead?> GetByIdAsync(int id);
+        Task<Lead?> GetByIdAsync(long id);
         Task<IEnumerable<Lead>> GetAllAsync();
         Task UpdateAsync(Lead lead);
     }
